step2 fill Monster class
step1 Main클래스에서 구현한 메소드를 일단 Monster클래스로 수정
Racing 클래스 만들면서 수정 해야 함
